Use supervisorctl to restart the thermalctl processes on pmon docker (#3411)
#### What is the motivation for this PR?

On pmon docker there are typically 2 thermalctl processes. One being the
parent process of the other. In order to restart the thermalctl process,
the existing code would assume that the process with the smallest pid is
the parent process. This is not always the case.

#### How did you do it?
A cleaner way is to user superivosorctl to restart the thermalctl processes
```
docker exec -i pmon bash -c 'supervisorctl restart thermalctl'
```
Modified the function "restart_thermal_control_daemon" in thermal_control_test_helpers.py with the above methodology to restart the thermalctl processes in the pmon docker