Add test for static route feature (#3429)
What is the motivation for this PR?
Test for the newly added static route feature.

How did you do it?
Add a vlan static route and send packets to an address in the prefix. Check if the packet gets forwarded to the expected port.

How did you verify/test it?
Run the test on kvm and dut and confirm that the test passes using images with static route feature.

Any platform specific information?
N/A

Supported testbed topology if it's a new test case?
This test case supports t0 topology.