Integrate the takeover notification step into the deploy script
This change integrates the takeover notification step into the deploy script.
This logic is contolled by a flag.
It was integrated into the deploy wrapper and not defined as a regression step, since it's much easier to control
the execution of this step by providing an argument to the deploy script. The send notification argument will be set
to True only from mars reg orch, where nightly regression instances are defined.

Change-Id: Id79503c4acfe2d9a03f7d24fd6b5ae6847525c16