[Sensor] Support Skip List of Sensors for a specific SONiC branch (#3701)
What is the motivation for this PR?
With different versions of SONiC images, due to the kernel driver version difference, the supported sensor or sensor attributes list could also be different. To make the sensor test case be adapted to this situation, extended the sensor_facts to support a skip sensor list on a specific SONiC image version.

A sensor skip list will be added to sku-sensors-data.yml which defined the sensors or sensor attributes that should be skipped in a specific version. Sensor_facts will parse it, if the DUT image version match, the set of sensors will be added to the skip list, a warning message will be printed out but the test case will not fail.

How did you do it?
Add a sensor skip list to sku-sensors-data.yml. e.g:

sensor_skip_per_version:
      201911:
        skip_list:
        - mp2975-i2c-5-62/PMIC-1 PSU 12V Rail Curr (in1)/curr1_alarm
        - mp2975-i2c-5-62/PMIC-1 PSU 12V Rail Pwr (in1)/power1_alarm
        - mp2975-i2c-5-64/PMIC-2 PSU 12V Rail Curr (in1)/curr1_alarm
        - mp2975-i2c-5-64/PMIC-2 PSU 12V Rail Pwr (in1)/power1_alarm
        - mp2975-i2c-5-66/PMIC-3 PSU 12V Rail Curr (in1)/curr1_alarm
        - mp2975-i2c-5-66/PMIC-3 PSU 12V Rail Pwr (in1)/power1_alarm
        - mp2975-i2c-5-6a/PMIC-4 PSU 12V Rail Curr (in1)/curr1_alarm
        - mp2975-i2c-5-6a/PMIC-4 PSU 12V Rail Pwr (in1)/power1_alarm
        - mp2975-i2c-5-6e/PMIC-5 PSU 12V Rail Curr (in1)/curr1_alarm
        - mp2975-i2c-5-6e/PMIC-5 PSU 12V Rail Pwr (in1)/power1_alarm
in the sensor_facts module, add a new function to get the SONiC image version, the new added list will also be parsed. If the image version match, the sensors will be added to the skip list. When checking the alarm, these sensors will be skipped but a warning message will be printed out.

10:29:59 test_sensors.test_sensors                L0050 WARNING| Sensor warnings:
[
    "sensor attributes [mp2975-i2c-5-62/PMIC-1 PSU 12V Rail Curr (in1)/curr1_alarm] is absent in version 201911", 
    "sensor attributes [mp2975-i2c-5-62/PMIC-1 PSU 12V Rail Pwr (in1)/power1_alarm] is absent in version 201911", 
    "sensor attributes [mp2975-i2c-5-64/PMIC-2 PSU 12V Rail Curr (in1)/curr1_alarm] is absent in version 201911", 
    "sensor attributes [mp2975-i2c-5-64/PMIC-2 PSU 12V Rail Pwr (in1)/power1_alarm] is absent in version 201911", 
    "sensor attributes [mp2975-i2c-5-66/PMIC-3 PSU 12V Rail Curr (in1)/curr1_alarm] is absent in version 201911", 
    "sensor attributes [mp2975-i2c-5-66/PMIC-3 PSU 12V Rail Pwr (in1)/power1_alarm] is absent in version 201911", 
    "sensor attributes [mp2975-i2c-5-6a/PMIC-4 PSU 12V Rail Curr (in1)/curr1_alarm] is absent in version 201911", 
    "sensor attributes [mp2975-i2c-5-6a/PMIC-4 PSU 12V Rail Pwr (in1)/power1_alarm] is absent in version 201911", 
    "sensor attributes [mp2975-i2c-5-6e/PMIC-5 PSU 12V Rail Curr (in1)/curr1_alarm] is absent in version 201911", 
    "sensor attributes [mp2975-i2c-5-6e/PMIC-5 PSU 12V Rail Pwr (in1)/power1_alarm] is absent in version 201911"
]

How did you verify/test it?
run sensor test with different versions of SONiC image.