Ignored test_everflow_dscp_with_policer on SPC2 and SPC3 due to HW Limitation
Need to adapt the test to hw limitation on SPC2 and SPC3
The Spectrum-2 and Spectrum-3 global policers, configured to PPS mode,
comply to 1% accuracy of the configured rate at time-periods
equal to 50,000 packets. As such, when configuring a global policer,
the policer needs to run 50000/CIR seconds to get to 1% error rate.

For example, configuring a policer CIR to 1,000,000 PPS will
comply to 1% accuracy at time-interval equal to 50 msec (50,000 packets).

[SONiC – Verification] Bug SW #2739322: [Functional] [everflow] | test_everflow_dscp_with_policer[cli-tor] SPC2 and SPC3 failed due to rate limit isn't accurate | Assignee: Roy S. | Status: Assigned

Change-Id: I364a4072094e650c6f8db181e90743a462e9ab5f