Beutify the ngts sub-directory
This change is a preliminary step, before enforcing pep8 rules in sonic-mgmt for the ngts path.
The change is modifying all the ngts sub-directory files to comply with pep8.
Any future change must comply with pep8, otherwise it will be blocked by the sonic-mgmt ci.

This change also adds new beutifier configuration files to be used by the sonic-mgmt ci.

Change-Id: Ica4c359314ee1ed476ae98f2005877409196bd62