Add root['XCVRD_LOG'] as one of ignored_item into the upgrade_allowed_diff.yaml
In 202012 'XCVRD_LOG' is added in the config_db.json, when the switch installed with 202012 image, with the pmon docker getting up, the xcvrd process in the pmon will get up, and xcvrd process will write the 'XCVRD_LOG' to the memory, after it is writen to the memory, if we do the save config, it will be saved to the config_db.json. But for different types of switch it takes different time to save the 'XCVRD_LOG' to the memory.so in upgrade_allowed_diff file, I add ignored_items, the item under it will be ignored when we compare the diff.

Change-Id: I97af69a9df5326e98f07984a96207eb5757fd6b6