Add root['KDUMP'] as one of ignored_item into the upgrade_allowed_diff.yaml
After upgrade the sonic image, during the image initialization, the hostcfgd service will be started, and during the
hostcfgd service starting, it will do "KdumpCfg load ...", and then sonic-kdump-config is called, at this time, the
'KDUMP' is written to the config_db.json file. when we copy the config_db.json file after the upgrade, we can not
make sure 'KDUMP' has been written to the config_db.json, so add root['KDUMP'] to the ignore item.

Change-Id: Iaa68c91fd066f71c2fbab330eb6d2c37c30aa606