Add tests test_acl_egress_drop and test_acl_drop on WJH
Change the way we parse the WJH table to support the second table that ACL gives (and Buffer in the future).
Change the verify_drop_on_wjh_raw_table and verify_drop_on_agg_wjh_table to only take the needed Channel.
Add verification for the ACL tests with verification for the Rules Info table

Change-Id: I6f97915c6d237efe94d358db7028d052669d3fcf