Exclude show syseeprom test from the build test
The ‘test_show_platform_syseeprom’ is verifying that the syseeprom values are the expected ones – for that we define expected syseeprom values per setup: onie version, cpld, etc.
Whenever a component version is change(onie, cpld) the setup info should be aligned as well, otherwise the test will fail.

Onie version on CI setup might be changed by users when debugging stuff manually. CPLD upgrades are also happening from time to time, and there are some time slots when the setup info is not aligned.

In order to reduce the false negative cases in the build test, I would like to suggest to remove this test from the build test(it will remain in the nightly regression).

Change-Id: I078f7f795572c5289baecc00590109f8a63257d3