Add workaround to not install app extension if it is already installed
There is a problem that wjh do not support uninstall.
Devops can not guarantee that each SONIC image will have wjh app
extension preinstalled, but on nightly regression would like to run
always with wjh. This workaround allows to use preinstalled app extension
to the image, and if image did not come with wjh, it will install
from script input parameter

Change-Id: I39317f54ac7b04e3e94501e4a105287cd43b2ec7