add new log error to ignore in lag test
Details:
This is an expected behavior as the hostcfgd indefintely listens for notifications from Config DB.

But because a reboot is issued in this case and the database container is killed before the hostcfgd,  we briefly see these messages in the Logs.

This log can be added to the ignore list.

Has not seen this before because, the hostcfgd was redesigned to use SubscriberStateTable which uses select underneath

See ticket, for more info: https://redmine.mellanox.com/issues/2835576

Change-Id: I220860cdf5fad08eb758a697c2cde934efe37810