Skip test_tunneling_between_sub_ports for Cisco-8000 (#4639)
Description of PR
Currently on cisco-8000 platforms, we only support uniform mode for DSCP. The subtests which are under “test_tunneling_between_sub_ports” are validating functionality using the DSCP Pipe mode and hence need to
be skipped.

Approach
What is the motivation for this PR?
DSCP Pipe mode support for IPinIP tunnels is currently unavailable

How did you do it?
Skip these tests for cisco-8000

How did you verify/test it?
Ran on a cisco-8000 device and verified tests are skipped instead of failing