modify script of apply configuration to not check docker status
Details:
the script is used to apply configuration on simx switch before configuring
connectivities, there is no need to check the docker status prior to apply
the configuration.
it also causes some stability issues on simx setups, the setup may be unstable before any configuration is applied but it is stable after the configuration is applied.

Change-Id: Ic3037798df7276748b3b316a6f20157b9913c423