Shutdown all bgp sessions before start the link flap test
The link flap test failed due to the oper_state of the interface will be change to UP after 30s.
This happens due to 3 facts: orchagent beeing single threaded, SPC1 has weaker CPU and SDK time to configure routes.
The port comes up in ~10 sec actually if you look at dmesg. But, when you shutdown the port orchagent starts to updated 6k ipv4 and 6k ipv6 routes with new ECMP wihtout the neighbor behind the port that is shutdown - this takes a while, now we startup the port, frr starts sending updates to orchagent and now we are adding back the neighbor (6k ipv4 and 6k ipv6 routes), this takes a while and only then the orchagent queue is free to process a notification that the port is up.

So shutdown the bgp sessions before this test to stablize the test
local fix for: https://github.com/Azure/sonic-mgmt/pull/4874

Change-Id: If1cacd2d6bf16d59c74ab49ae3aa6e8e88f3f31e