Send regression takeover notification both to sonic-mgmt and the dut
In some cases, the users will be logged in only to the sonic-mgmt cont. and not to the dut, thus they won't receive any notification about regression takeover.
This commit adds support to notify both dut and sonic-mgmt logged-in users. If the script won't be able to connect one of the players- it would silently fail,
since at this stage in the regression we cannot guarantee that the dut and sonic-mgmt are alive.

Change-Id: I9a1983a9615ddbdd4a3c71839b15e22338848139