change the p4 examples fixture scope
When the scope of fixture is session, the teardown of the fixture will be done only when all the test case finished in the session.
The scope of p4_examples_config fixture should be set to package instead of session, since the config should be used only for the test case in the p4 examples package, else if it is set to session, when run the p4 examples test case together with the other test case especially vxlan test case, the vnet config added for the p4 examples will effect the vxlan test case, since for SONiC, only on vnet can be added.

Change-Id: I6d4bcd5d3355ae3d1a7a8dd4f27df63b7f465ecf