Align the wjh buffer test with the dynamic buffer configuration
When a packet is received the following steps will be performed:
1.	Extract the DSCP value from the IP packet.
2.	Map the DSCP value to traffic class according to DSCP_TO_TC_MAP.
3.	Map the TC to QUEUE according to TC_TO_QUEUE_MAP.

Originally, the test worked on top of SDK default configuration which maps all the DSCP to TC zero and then to QUEUE zero.
Now that we loaded the QoS configuration from SONiC, the wjh test should expect the traffic to arrive to QUEUE one.
The test was adjusted to get the expected queue ID based on the buffer comfiguration mode. In case of static- it will always be zero,
while in case of dynamic the specific traffic used in this test will be mapped to queue 1.

Change-Id: Icc236cf198b4d93c8895b27fe2df1081089ce1c5