Fix issues that cause failure on BSL regression
Details:
1) generic_sonic_output_parser may return interface ips as a list,
if interface has more than 1 ip, or as a a string if interface has only 1 ip.
In that case, if ip_addr = interface_dict[ip_key] is list, then ip_addr.split('/')
result in error - therefore the added fix.
2) use cli_object existing functions in shutdown_interface.py script,
it's better practice to call the functions from the cli_object, not the class.

Change-Id: I02bbb0a95e80bf90a24abc4ed18057adfee25823