Ignore ansible errors when executing shell command test_interface_binding
This change is to enhance the stability of test_interface_binding. There is already logic
to verify if the dhcp relay agents are running correctly in test_interface_binding by
executing command "docker exec -it dhcp_relay ss -nlp | grep dhcrelay" on the DUT. And if
this command get nothing from the DUT, the test will try to reload config to make sure
dhcp relay agents are good before running other tests.
However, if "docker exec -it dhcp_relay ss -nlp | grep dhcrelay" does get nothing(stdout is
empty ), the ansible method will fail immediately and won't return to the caller function,
so the verify and config reload logic will be useless, which means the effort to make this
test stable is in vain.
To resolve this, need to set module_ignore_errors as True to ignore ansible errors.

Change-Id: I2645700e4f940dcdce68c1991ee6b7394902c4de