Save config_db after configuration cleanup for test_dhcp_relay
In test_dhcp_relay, vlan/IP/dhcp relay configurations are saved to db in set up phase. But when doing config cleanup, the ip and vlan configurations are only deleted but not saved. So any other tests with a config reload action after dhcp relay test may be affected and fail.
So add config saving after all configurations are cleaned in teardown.

Change-Id: Id9f1864885209f33faf2b209178469af0c483881