Update script to check dockers are up after configuring switch
Details:
The following script is used in the bringup_containers script to configure the simx switch,
before setting the connectivities of the setup.
In case dockers on the switch are failing after configuration,
the split tap devices won't be created and the bring up containers will fail on connection configuration,
when in fact the real issue is that the dockers on the switch are down.
this way the script will give true indication for what is the issue.

Change-Id: I0e681674a4f6ba01d79d9709fc993ff161e6a006