Change the filter for the sniff used in p4 sampling test cases
There should be some issue in the sniff, when the pri of the vlan is 0, filter 'vlan 40 and src 50.0.0.1 and dst 50.0.0.2' works as expected, but when the pri of vlan is 1, then 'vlan 40 and src 50.0.0.1 and dst 50.0.0.2' can not work success. so change it to only use vlan instead of filter vlan id, and also add ether src, ether dst in the filter.

Change-Id: I35210ccc2a9478cf6a58a3de12ff5eaf47e7a174