change apply_basic_config reboot checks
Details:
apply_basic_config should check only if dockers are up after reboot,
and not if ports are up as well. this is because this function is also
used to configure SIMX switches before connectivity is configured.
if port status will be checked it will fail on a simx setup.

Change-Id: If59b17965699c4191cce2c28bd3738d60eabc7d5