Remove one of the ignored err msg
The err msg "ERR swss#orchagent: :- addOperation: Vxlan tunnel 'neighbor_advertiser' is already exists" which is reported in the #2997615 together with the other err msgs.since #2997615 is not a bug, the err msg is added to the ignore list, but for Vivek's closer look, another ticket is raised for this err msg: 3031477.

Vivek want to check if the test case will still hit this err msg in the regression. so just remove it. and if we verified that the err msg is hit, need to add err msg to dynamic ignore list

Change-Id: I9b0e1204cc99d4eba1c4877b191af33b9075641e