Update apply_configuration script for simx setups
Details:

this script is used to configure the simx setup during the containers bringup stage.

in cases of SPC4 simx setups, the platform and machine attributes cannot be configured correctly
on the Onie vm on the simx docker.

this is a workaround so that the simx SPC setups could be configured correctly.

Change-Id: Ieed75e26f4661046bc44f8dfd681f033e3317a30