Apply fix for bug #3028429
Details:
Bug info:
[SONiC – Verification] Bug SW #3028429: [Functional] [LAG] [SCALE] | members are in unexpected state on Lag with max members | Assignee: Sharon Lutati | Status: Assigned

Root cause:
The test fails because of teamd LACP port state flapping due
to a lot of LAG members are configured with the loopback cables.
The W/A is to disable the LAG members which are connected with the loopback cables.

The fix is to apply the workaround of disable the LAG members which are connected as loopback.

Change-Id: Ifda2dc08e70103087dfba4ae5fe3e92dcfb00872