Update script to check that all ports exist after configuration is applied
Details:

this script is used to configure the simx setup during the containers bringup stage.

in cases of simx setups, if not all ports (including split ports) exist on DUT
the connectivity of the setup will not be configured correctly.

It is more correct that the script will alert that the ports were not configured correctly,
than connectivity configuration will fail over port doesn't exist.

Change-Id: I721beb5825e0ed1efa9e587d9763509de3f725d1