Update DPB tests cleanup
Details:

DPB tests should apply basic setup configuration after their run because
some DPB bugs could cause ports to be unmapped which in that case the tests
cleanup is not enought to return the switch to the previouse state.

Qos and dynamic buffer configuration is already included in the apply_basic_config.

The DPB feature is not enabled on 202012.

Change-Id: Id9d12aef13d113d97dd40a3a194488f500dfde20