Remove dafault acl tables before test subinterfaces based on portchannel
In sub interfaces test, the portchannel cases add ethernet ports to portchannels, but these ports are already in the default acl tables, which can not be added to a port channel.
Remove the default acl tables before the port channel cases, and the will be added back aotomatically by config reload in teardown.

Change-Id: Ic78917bb8c2eb25861a4315d4ba00065ec422747