Add new test for QoS new cli command
Details:
This tests checks the functionality of CLI command "sudo config qos reload --ports",
the documentation for this command can be found at:
https://wikinox.mellanox.com/pages/viewpage.action?spaceKey=SW&title=Design+-+config+qos+reload+--ports

the test removes qos and buffers configuration from random ports,
and the checks that the qos and buffers configuration is restored after
"sudo config qos reload --ports" is configured.

for the implemantation some fixtures were moved to tests/conftest.py,
so it could be used by the test.
some imports were changed because of that.

Change-Id: I0b1a1921a7737fa5d495c4f76546e93d19fce5fa