Update tests/logs skips for DPB/COPP tests
Details:
bug #2960190 is not affecting all DPB tests, only several tests should be ignored
over this bug.

bug #2934289 is no longer observed, and it is a logging issue thatr should be addressed in dynamic
ignores on logs and not test.

bug #3006650 is closed.

bug #2980204 is a verification issue that affects simx only.

DPB cleanup should be in session scope and not on moudle,
because in moudle scope it will be called after tests on each python file, not on all the tests.

Change-Id: I066ea5b18a0f8a45228090105c4dbaf07e12cc1a