Update DPB tests to run on 4600/C platforms
Details:
Modified the DPB tests to calculate the ports created after a breakout from the output of
the breakout command instead of using information based on parsing of the platform.json file.
Because on platform 4600/c the port interfaces and lanes are not synced,
so port interface name after breakout can not be concluded from lanes number.

This change allowed basic test test_dpb_on_all_ports to run Succesful on 4600/C platforms.

Other tests that need to be modified are for now disabled until
the necessary changes to the tests could be tested, see verification bug for more info:
[SONiC – Verification] Bug SW #3103647: [Non-Functional ] [DPB] [4600/C] |
 adapt DPB tests to platformS 4600/c | Assignee: Sharon Lutati | Status: Assigned

Change-Id: I3cd1bc6e6d7e64cccaec9b8863cab56db701a95c