Fix for bug #3102634
Details:
Fix is to compare only the relevant keys that should be updated by
the "config qos reload --ports" commands, instead of comparing the entire
config_db.json.

This is because keys like FLEX counters could take several minutes to update in the config_db.json,
and they are irrelevant to the check of "config qos reload --ports" configuration.

Change-Id: Ib8aa7d6b8b64c5840f3c13924433a116ec8e492e