Workaround for issue - https://github.com/Azure/sonic-buildimage/issues/11216
Details:
Because of the issue https://github.com/Azure/sonic-buildimage/issues/11216,
the release key is 'none' on image 202205, in that case, the branch is set to be 'master'
by default and not 202205.
This means that the log ignore will not be applied on the 202205 branch as expected.
so the workaround is to remove the branch - so the build could pass.

Change-Id: Ie5f04d44a0612fad6b0805339fa60ffe0e36d201