Fix host vlan test for RM#3047936
The test need to configure a new mac address other than the switch mac on a vlan interface. The community uses "00:11:22:33:44:55" as the new address. But for our asics, we don't support this kind of mac address(see the rm ticket for detail). So add a new logic to get a valid mac address for Mellanox duts.

Change-Id: Ic115b8f3bb1a9cf3fc2eb34d27b6352997c86315