Skip test_ram_usage on ASAN image
Details:

It is expected that ASAN image will consume more RAM than
regular sonic image, no need to monitor spicel values for that image.

The skip is with fixture and not a part of tests_to_be_skipped_conditionally.yaml,
because currently there is no support for dynamic skips on asan image.

Change-Id: I720f4117b0168e0b30a9e9a5caf5044bf80e1cf8