Fix subinterface test issue for RM#3117697
The root cause of RM#3117697 is dict is used to store the dut ports and ptf ports selected for the test, but since there's no order in python dict, the dut port could be paired with a wrong ptf port(the dut ports is returned in dict but the ptf ports is returned in list of the dict values, and they are zipped in the caller to do iteration). So use OrderedDict instead of normal dict to store the selected dut ports and ptf port in get_ports.

Change-Id: I0e0e2d42bac93ab04e003848765c217e33b26c75