Reduce runtime in deploy
The first step after installation of image was disable ZTP.
Later uploading the config files and reboot with validation of dockers.
ZTP tooks ~2.5-3 minutes.
To reduce the runtime, the first step will be upload the config files.
The new config files doesn't includes the ZTP. And if yes, the disabling will be during dockers bring-up.

Currently this change reduce ~3-4 minutes of runtime.


Change-Id: I7a96ecfd04f66fbc0546efb9ea2755a9ab504ebc
Signed-off-by: Anton <antonh@nvidia.com>