Fix copp issue on r-tigon-11
Becuase the perfomance of test server of r-tigon-11 is lower, so the send packet rate is low than case expected. As the test desgin, we can increase the send packet rate by setting the parameter of send_rate_limit. The default is 2000, now it is changed to 4000 which is selected by test.

Change-Id: If79bdfa9098f1134065b50dabd5b081f4f0dd1db