Restore the setting timezone logic in deploy image
The logic should not be moved with ZTP relevant codes in commit http://10.7.77.140:8080/c/switchx/sonic/sonic-mgmt/+/115990.
Because the apply_basic_config function is not called when deploying community setups, so the timezone will not be set.
Move the logic back to deploy_image function which is for both community and canonical deployment.

Change-Id: Iaf0faa6341528527959df5cfae435ecd937a59b0