Update the regexp in test_check_existing_network_show_commands
The regexp in validate_ip_bgp_network function is to match the bgp network of the default route:

*= 0.0.0.0/0        20.0.0.2                 0             0 501 i
*>                  30.0.0.2                 0             0 501 i

'>' indicates the best route. The old expression matches the output when the best route is via 30.0.0.2, but sometimes the best route is via 20.0.0.2, and the test will fail.
Update the expression to match both.

Change-Id: Ice39898500e68379cd77fc6dfc76ba8274f3537d