Update ignored error list for RM#2856406
These 2 tickets are duplicates of RM#2856406:
[SONIC - Design] Bug SW #3179716: [Non-Functional ] [Error in log] | ERR kernel: [ 382.187849] mlxsw_minimal 2-0048: Failed to query PWM | Assignee: Cong Hou | Status: Rejected
[SONIC - Design] Bug SW #3177342: [Non-Functional ] [platform] [error in log] | ERR kernel: [ 345.246800] hwmon hwmon13: Failed to query module temperature | Assignee: Cong Hou | Status: Rejected

The root cause of these errors are the same:
While shutting down the pmon, it will kill all the running daemons, if some transaction(access some sensors via sysfs) is still not finished, errors msg could print out. The error msg could be various depends on which sensor is being accessed, but they are all sharing the same root cause.

Add the new errors to the ignore list of RM#2856406

Change-Id: I57bdec6e0c5e3460ea3a3667d5acc85e48a1d7ce