Add compatibility for dualtor setup when getting topology from noga
The lower tor dut in a dualtor setup uses 'dut-b' as the description in noga, and which is used by the topology object as the key of the dut object. Need to update the logic for getting topology from noga to be compatible with dualtor setups, otherwise all codes use topology.players['dut'] will get key error.

Change-Id: Iec6441fd575afa2cf97dcf5589f54689d4505793