Added a test case for fast reboot from other vendor NOS to SONiC.
Details:
As part of MSFT FR, a new test case was added to fast reboot script.
This test case mimics a fast reboot from other vendor nos to SONiC, by flushing all DBs that
are part of the boot process, so the boot will be initiated with clean DBs as if it was booting from other nos.

Change-Id: I5534dad52ead5f021f8bfb5d867510628e48adf0