Fix auto negotiation port toggle test
When generating the default config, use only the min speed instead of random speed. Because higher speed may cause issue in the scenario when dut is AN and host peer is forced speed which is in the port toggle test flow.
The test was originally designed to use min speed but later changed by an improvement which is not necessary here, the min speed is enough for the test purpose.

Change-Id: I13adf9e4bf7199282a77039a05ea193bb0d03df3