Add start bgp step in bgp bounce test
In the bgp bounce test case, template is used to update the bgp config. and it is required to have 'admin_status' in the BGP_NEIGHBOR of configDB. Otherwise an error will be thrown and the peer may not be updated:
ERR bgp#bgpcfgd: Peer '(default|fc00::5e)': Can't update the peer. Only 'admin_status' attribute is supported
The initial config of community setups doesn't contain 'admin_status', so need to configure it before the test.

Change-Id: I8616d047ec0e25f1b74d2e7807f7b68baba5da2f