Add qos remapping test cases for Nvidia only
Add 2 test cases for Nvidia platforms:
1. test_active_tor_normal_traffic
The reason to add this one is because our remapping is per port not per tunnel, need to verify the normal traffic is not affected by remapping.
2. test_traffic_between_servers
The reason to add this one is this scenario is not covered by community tests.
Will contribute these 2 cases to community when they are stable.

Change-Id: I3e3a7aace0d4a1567dc05a28b8fa3cc67346bb16