Increase the wait time in nexthop and neighbor test
This is to fix issue [SONiC – Verification] Bug SW #3205851: [Functional] ["CRM"] | *available" counter was not incremented
When the nexthop/neighbor is removed there will be 2 steps to fully release the resources.
First, almost immediately, the resource occupied by neighbor and nexthop will be released.
And then after some time(may be more than 30s on some platforms), the resource occupied by the SDK/SAI internal object will be cleaned up.
So increase the wait_until timeout to 60s.
And updated the script to pass the pre-commit check.

Change-Id: Id7489690868b4e0f11986cd5276d1e72617f41dc