Change date format in filename of sanitizer dump
Details:
If the archive file name includes a colon (‘:’), then it is assumed to be a file on another machine.
This naming convention causing the "tar -xzvf " to fail on our ASAN dumps.

Name according to old format:
dump_nightly_sanitizer_files_08_19_2022_06:39:33.tar.gz

Name with new format will be, for example:
dump_nightly_sanitizer_files_08_19_2022_06-39-33.tar.gz

It is also the date naming format in the ASAN dump itself, as you can see here:
./2022-08-19_06-37-42_2022-08-19_06-37-42_wjhd-asan.log.22

Change-Id: I4178aa0117255c39e1be4d94c2f09008a9972a0a