Add a test case to cover issue RM#3322173
This test is to verify the issue in RM#3322173.
When BUM(broadcast/unknown unicast/multicast) traffic is received in a vlan, whether it should be dropped or flooded
in the vlan depends on the priority. Lossless traffic should be discarded, while lossy traffic should be
flooded to the other ports in the vlan.

Change-Id: I90e0a74897b1293e81c0fe1dc9c29aaf7d60ee2c