[fast/warm-boot test] Fix the misleading print of zero traffic loss and test pass when the traffic is not restoring
In the warm/fast-boot test, a certain amount of packet is sent(about 50k), the date plane loss is derived by the number of consecutive packets lost during the reboot. However, if the traffic won't restore, the loss time will be calculated as 0, and the test will pass(not expected). This change is fixing the case, so if the traffic won't restore, the test will fail and a meaningful message will be printed to the user.

Change-Id: Ie78a7d4b03640cd95eedd30db5ee30df6984d030