Improve the config check reload
Previously community added a new fixture to check core dumps and compare the config_db.json before and after the test. If there is core dump or config change during the test, a force reload will be performed at the teardown phase.
However, sometimes the force reload is executed when the system is not ready(some test will do reload/reboot), in this case, the force reload could cause some problems and the system may take very long time to restore, which will affect the following tests.
I add a new logic in the config_reload method to support wait before performing the force reload, and use this logic in the core_dump_and_config_check.

Change-Id: Ib2671090db49b321024c3dce4e48c15a7dddc23e