Do not enlarge PFC WD polling timer if it is shorter in QoS testPfcStormWithSharedHeadroomOccupancy
The PFC WD should be enabled before the test. There is a dependency between PFC WD timers:
The detection, and/or restore timers of each port should not be less than the polling timer. By default, they share the same value.
This means we should not enlarge the polling timer if PFC WD has been enabled. Otherwise, the dependency can be violated.
But we should shorten the existing polling timer if it is greater than the test parameter because all the timers share the same value in the test parameters.
Otherwise, the detection and restore timers will be greater than the polling timer, which also violates the rule.

Signed-off-by: Stephen Sun <stephens@nvidia.com>
Change-Id: Icad3f1532fc1585aa5d40d8ac7583bf80782c3bf