Changed the "simx_uncovered" pytest mark to "physical_coverage"
Details:
The new mark, will declare the tests that are being covered on physical setups during CI/build, and not the features that Simx doesn't cover.
It will enable running tests on CI/build on Simx and Physical together, and to test different functionalities.
In addition, added 2 more tests: Vlan and ECMP calculator to be covered during CI/build on physical setups.

Change-Id: I55215b5c9da858967770f9acf9aaf96c946591b8