Fix issue in testPfcStormWithSharedHeadroomOccupancy
Fix issue when the polling timer is recovered at the end of test

```
RunAnsibleModuleFail: run module command failed, Ansible Results =>
{"changed": true, "cmd": ["pfcwd", "interval", "800"], "delta": "0:00:00.614236", "end": "2023-03-20 04:03:39.275514", "failed": true, "msg": "non-zero return code", "rc": 1, "start": "2023-03-20 04:03:38.661278", "stderr": "unable to use polling interval = 800ms, value is bigger than one of the configured detection time values, please choose a smaller polling_interval", "stderr_lines": ["unable to use polling interval = 800ms, value is bigger than one of the configured detection time values, please choose a smaller polling_interval"], "stdout": "", "stdout_lines": []}
```

We only reconfigure PFC WD polling timer when it is less than the existing value.
When we recover the polling timer, we are enlarging it, which causes an error
because we make the polling timer less than the detection timer, restoration timer
who are the same value as the current polling timer.

Fix
Restore, which is to enlarge, the detection timer and the restoration timer first and then the polling timer.
The previous detection timer and the restoration timer are the same as the polling timer to be recovered.

Signed-off-by: Stephen Sun <stephens@nvidia.com>
Change-Id: Ie6a277a86d36a73042e5e1d40c407541d10c90e4