Upstream PR #8050
Details:
The policer can confirm test deviation for Nvidia vendor, only with min number of packets of 50,000.
As the number of packets is determined in ptf test/test only as a result of test send_time,
after testing on Nvidia platforms minimal send_time which was stable and achieved send packets in 50k range was 75 sec.

Change-Id: I0eb1cb498afdf14d66f9b6973148f7f83f780f5b