Use wait in the checks in route flow counters test (#8052)
The test statistically failed on some of our platforms with 202211 image, and the cause was that there was a small delay between the static route is added/deleted and the output of route flow counters were updated. The delay was very small(less than 1s) but it may cause the test to fail sometimes.
So, it would be better to use wait_until with the checks to improve the stability of the test.
And I've replaced the logger.info with allure.step in tests.common.plugins.allure_wrapper to clean up the redundant text in the test script. It also prints entries in the test log.

What is the motivation for this PR?
To improve the stability of the test.

How did you do it?
Use wait_until with the checks
Use allure wrapper and remove the logger info