[tests/configlet] Remove -n in apply-patch in add_rack test (#6146)
Description of PR
Summary: Remove -n in apply-patch of test_add_rack.

What is the motivation for this PR?
The -n was added because not all of YANG models are present in SONiC. Now the YANG model is complete. So we can remove the -n option.

How did you do it?
Remove -n in apply-patch operation.

How did you verify/test it?
E2E test.