Reverted VNET Vxlan test params to prevent route configuration issues
Details:
Vnet Vxlan test fails with the current params: 16K routes, 1 vnets, and 4K endpoints.
The test fails due to a combination of ipv4 and ipv6 tunnels, due to ipv6 parameter enabled on sonic-mgmt branch.
The ipv6 is disabled on 202205 branch, and the number of vnets is 1 instead of 8 in the 202211 branch.
The PR reverts the parameters back to the 202205 values, the test passes with those parameters.

Change-Id: Ic69d531b8e80612e53fe92e3dd5818488918fe10