Update ignore on fw stuck dump test
Details:
This test is causing the crash of the syncd and leads to undefined behavior of ASAN.
When the application is crashing cleanup of used memory is not involved so the leaks reported by ASAN are not real leaks.
This is why test should not run on ASAN image.
See more info on: https://redmine.mellanox.com/issues/3481446

Change-Id: I1d5acedde0ee10d9212595c373de954aef4e5701