[Snappi] Skip warm and fast reboot RDMA test cases for Cisco devices (#8415)
Summary: Cisco devices do not support warm or fast reboot operations. If such an operation is performed, it leaves the switch in a bad state, resulting in failures. This change skips all cases with warm and fast reboots for cisco devices.

What is the motivation for this PR?
Lots of failures related to ARP were seen with warm and fast reboot RDMA cases in cisco switches. Cisco switches currently do not support warm and fast reboots, so these failures are expected. Hence, these test cases will be skipped if it is a cisco switch,

How did you do it?
Check device's asic type and skip the appropriate cases.

How did you verify/test it?
Test cases are skipped on lab switch.

Any platform specific information?
Cisco-8000