fix test_cacl_function
When checking ssh port is stopped, the delay for wait_for is 30s, which is too long. In the 40s timeout, it can be tried only 2 times at most
But it will take more than 60s to get the results. In the test, it will sleep 120s, and then clear the alc rule. It means that we need to run get_snmp_facts before 120s. Before get_snmp_facts, besides checking ssh port is stopped, we also need to check ssh port is not started. So, when checking ssh port being stopped takes much time, it will cause case to fail. So, Changing 30 to 10 will accelerate the return response.

Change-Id: Ieab5b1c2056b22076cd2302a2dfd16250d6e6a65