Added init flag to docker run command, to handle zombie ssh sessions.
Details:
Docker does not run processes under a special init process that properly reaps child processes, so that it is possible for the container to end up with zombie processes that will eventually prevent creating new processes when it will exceed the maximum processes allowed.
The fix adds to docker run command an init flag that will run an init inside the container that forwards signals and reaps processes.

Change-Id: I15c4e880368b6713fad42883af3601322a27e403