Added PLL (Phase-Locked Loop) Event test case.
Details:
This PR adds a PLL lock event scenario to the mellanox sdk dump test, and simulates a PLL lock failure from SDK side.
The PLL lock event is added as a pytest parameter, and expects errors in log due to this event.
The trigger script uses a unique event ID given from SDK, I added an argument which receives this id from the test script.

Change-Id: Id0410b8e4c7005ff185165ebecb174a0edc778cf