Remove old cpld of bulldog
The fwutil CPLD update test is skipped on bulldog due to a bug RM#2382365. If the old CPLD is installed by the test, it cannot be restored without hw team's help. And when there is RM issue, the test could accidentally run on bulldog. To avoid this, remove the old CPLD in the firmware.json file.

Change-Id: I29c053e488e3c4291f2f75d5b0b762c672ab2a5a