Optimize the run time of dynamic port breakout test
When check the transceiver presence state, no need to call the cli command 1 by 1 for every port, we can call the cli for once to get the information for all the ports, it could save the run time of the test case. After the change, it could save around 18 minutes for all dynamic port breakout test.

Change-Id: I0180f8d3ab19aa8578600c8d44a7a9d8488745b5