Add new egress drop cases
Add two new test cases for Nvidia platforms to cover the test gap introduced the deviation of the dualtor acl rule.
Nvidia platforms use the egress drop rule instead of the ingress rule to drop the upstream traffic received by the standby ports.
Comparing to the ingress rule, the egress rule should be applied on all ports but not only the vlan ports.

Change-Id: Ibb86ba8077cc526c64bf234dc165907391660779