Fix test_config_interface_speed issue
For nvidia device,the headroom size is related to the speed and cable length. When platform is x86_64-nvidia_sn5600-r0 and above,we need to choose interface whose cable length is 40m not 300m. Because this platform supports speeds of 400G and above, if we use 300m cable length it will exceed the headroom limit and cause some log errors.

Change-Id: Ie6479346b019e78c6fe95ad7cc13e03d734d601f