Fixed issue in watchdog API test.
Details:
The test activates watchdog timer with 10 seconds, then waits 2 seconds, fetches the remaining time (8 seconds), and then activates again the with 10 sec and immediately gets remaining time again. The test expects that the new remaining time is about 10 seconds that is larger then remaining time of 8 seconds, since we did not wait anymore.
Turns out that there is a delay of about 1.05 seconds until "get_remaining_time" is called. So the API returns 8 sec, probably rounds to closest lower value.
So the test did not execute second get_remaining_time quick enough after arming the watchdog.
This fix extends the waiting time by 2 seconds, to catch the delay until the function of getting the remaining time is called.

Change-Id: I9a9da5dbb443f348ae774e609d3a4308b21d85f2