Fix test issue in fast reboot with max port
For lionfish setup, the physical ports have different number of lanes, when the loopback ports don't have same number of port lanes, should avoid to split the ports.

For example, port48 and port49 are connected as loopback, but the port48 has only 1 lane, and port49 has 4 lanes, if split port49 to 4 ports, then 3 of the ports are in down status, it will fail the test case.

Change-Id: Ie101ede3a3e5a2fa6fc9f3cdf45509dda706e8a5