Save platform params in new json file
Sonic only.
The hwsku can be changed by image in deploy.
To avoid failures of deploy and tests need to update the stored in Noga platform params.
But CI setups are read-only.
So the params will be stored in the new json file device_platform_info.
Script trying to read from json file, if not successful - read from Noga and update json file.
In deploy, after image installation, check the current hwsku, and if needed, update the json file again.
The updating of json file will be a part of setup bring up flow.


Change-Id: I6d7ac08a6d6789d55b1e6caa8ffd28bb79e0a028
Signed-off-by: AntonHryshchuk <antonh@nvidia.com>