Cherry pick of http://10.7.77.140:8080/c/switchx/sonic/sonic-mgmt/+/135772
Details:
Fixed log handling issues on fast/warm reboot:
Fixed reboot type source fetch logic: Previously, the reboot type source was based on the test name. However, it doesn't work with the upgrade path test, as it does not include the reboot type in the name of the test. This fix treats the mentioned case by getting the name from the upgrade type parameter.
Change-Id: Ia6ed297e83fdf52b162b7a960ea2e94e26a903a9