Added a logic in which the removal of config db file occurs only when installing a base version with sonic installer.
Details:
Previously, the removal of config db happened each time we use "install_sonic" method.
This method is being used both on installation of base version and target version.
The removal of config db after installing the target version, caused the following issue:
The switch performs a cold boot instead of fast boot, due to minigraph deployment. Further details are provided in RM #3571214
This PR sets the removal of old config db to occur only after installing a base version of sonic and loading clean minigraph.

On the way, fixed a typo in the variable name of target version, it installs the base version, therefore the name should be base version.

Change-Id: I0d1860765aa28ad25450fee3ad94ed8121186d76